+++ web/app/repositories/index.js
@@ -1,5 +1,0 @@
-import * as users from './users.repository';
-
-export {
-  users
-};