+++ web/app/store/actions.js
@@ -1,9 +1,0 @@
-import * as menu from './actions/menu.actions';
-import * as preventDoubleClick from './actions/prevent-double-click.actions';
-import * as users from './actions/users.actions';
-
-export {
-  menu,
-  preventDoubleClick,
-  users
-};