+++ web/app/store/actions/menu.actions.js
@@ -5,6 +5,6 @@ let setActive = (menuName) => {
   };
 };
 
-export {
+export default {
   setActive
 };