+++ web/app/store/actions/prevent-double-click.actions.js
@@ -10,7 +10,7 @@ let disable = () => {
   };
 };
 
-export {
+export default {
   enable,
   disable
 };