+++ web/app/store/actions/users.actions.js
@@ -12,7 +12,7 @@ let clearActive = () => {
 };
 
 
-export {
+export default {
   setActive,
   clearActive
 };