+++ test/specs/react/components/application.component.spec.jsx
@@ -1,5 +1,5 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import React from 'react';
+import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {Application, Header} from '../../../../web/app/react/components';
 import store from '../../../../web/app/store/store';