+++ web/app/application.jsx
@@ -5,3 +5,10 @@ import './styles/main.scss';
 
 //application bootstrap
 import './router.jsx';
+
+//NOTE: lets re-throw error by default
+import {Promise} from 'bluebird';
+
+Promise.onPossiblyUnhandledRejection(function(error){
+  throw error;
+});