+++ web/app/mock/data/users.js
@@ -1,18 +1,13 @@
-let oneTwoThree = {
+export const oneTwoThree = {
   id: 123,
   firstName: 'Test',
   lastName: 'User',
   username: 'test.user'
 };
 
-let oneTwoFour = {
+export const oneTwoFour = {
   id: 124,
   firstName: 'Test2',
   lastName: 'User2',
   username: 'test2.user2'
-};
-
-export {
-  oneTwoThree,
-  oneTwoFour
 };