+++ web/app/pages/desktop/router.hooks.js
@@ -1,15 +1,15 @@
 import store from '../../store/store';
-import usersActions from '../../store/actions/users.actions';
+import usersActions from '../../store/users/users.actions';
 import usersRepository from '../../repositories/users.repository';
 
-let desktop = {
+export const desktop = {
   onEnter: function(nextState, replace, callback) {
     store.dispatch(usersActions.clearActive());
     callback();
   }
 };
 
-let withResolves = {
+export const withResolves = {
   onEnter: function(nextState, replace, callback) {
     usersRepository.getUser(124).then(function(user) {
       store.dispatch(usersActions.setActive(user));
@@ -24,9 +24,4 @@ let withResolves = {
       callback();
     });
   }
-};
-
-export {
-  desktop,
-  withResolves
 };