+++ web/app/store/prevent-double-click/prevent-double-click.actions.js
@@ -1,12 +1,14 @@
+import * as constants from './prevent-double-click.constants';
+
 let enable = () => {
   return {
-    type: 'PreventDoubleClick::enable'
+    type: constants.ENABLE
   };
 };
 
 let disable = () => {
   return {
-    type: 'PreventDoubleClick::disable'
+    type: constants.DISABLE
   };
 };