+++ web/app/store/prevent-double-click/prevent-double-click.constants.js
@@ -1,0 +1,2 @@
+export const ENABLE = 'PreventDoubleClick::enable';
+export const DISABLE = 'PreventDoubleClick::disable';