+++ web/app/store/users/users.constants.js
@@ -1,0 +1,2 @@
+export const SET_ACTIVE = 'Users::setActive';
+export const CLEAR = 'Users::clear';