+++ web/app/routes.jsx
@@ -3,7 +3,8 @@ import {
   browserHistory,
   IndexRoute,
   Route,
-  Router
+  Router,
+  applyRouterMiddleware
 } from 'react-router';
 import {Provider} from 'react-redux';
 import {syncHistoryWithStore} from 'react-router-redux';
@@ -12,12 +13,16 @@ import Application from './react/components/application.component.jsx';
 import NotFoundPage from './react/components/not-found.page.jsx';
 import DesktopPage from './pages/desktop/desktop.page.jsx';
 import {routes as desktopRoutes} from './pages/desktop/module.jsx';
+import useScroll from 'react-router-scroll';
 
 let history = syncHistoryWithStore(browserHistory, store);
 
 export default (
   <Provider store={store}>
-    <Router history={history}>
+    <Router
+      history={history}
+      render={applyRouterMiddleware(useScroll())}
+    >
       <Route path="/" component={Application}>
         <IndexRoute component={DesktopPage} />
         {desktopRoutes}