+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # master
 
+- source maps were not loading properly with just the -d flag so modified webpack configuration file
 - minor code refactoring
 
 # 0.11.0